feat(rocm): add rocm_compat shims and cudarc-hip driver symbols for decode-graph
Extend the ROCm compat layer and the cudarc-hip fork so the mistral.rs
decode-graph/GDN/sort/top-k kernels build under hipcc:

- rocm_compat/cuda_pipeline.h: map CUDA <cuda_pipeline.h> __pipeline_* onto
  the gfx1250 async global->LDS builtin (__builtin_amdgcn_global_load_async_to_lds_b128
  + s_wait_asynccnt) with a synchronous fallback for targets without it (gfx1151).
- rocm_compat/cub/block/block_radix_sort.cuh: alias cub::BlockRadixSort to hipcub.
- rocm_compat/cuda_bf16.h: add __floats2bfloat162_rn (pack two floats into bf162).
- rocm_compat/cuda_runtime.h: add cudaDeviceGetAttribute, cudaDevAttr* constants,
  occupancy/cooperative-launch, and cudaMemcpy2DAsync mappings.
- cudarc-hip driver: add PinnedHostSlice + CudaContext::alloc_pinned,
  CudaEvent::elapsed_ms, the CUresult/CUmemPool_attribute enums, mempool/graph-mem
  driver fns (cuMemPool*/cuDeviceGraphMem*), cuMemcpyHtoDAsync_v2 wrapper, and
  cuMemHostAlloc/Free + cuEventElapsedTime.

+++ cudarc-hip/src/curand.rs
@@ -205,7 +205,7 @@ impl CudaRng {
     pub fn new(seed: u64, stream: Arc<CudaStream>) -> Result<Self, CurandError> {
         let ctx = &stream.inner.ctx;
         if let Err(e) = ctx.bind_to_thread() {
-            return Err(CurandError(e.0));
+            return Err(CurandError(e.0 as i32));
         }
         let gen = unsafe { result::create_generator()? };
         unsafe { result::set_stream(gen, stream.inner.cu_stream)? };
